feature(refactor): rename files and edit the paths where they are being used
Finishes #149631313
