+++ .gitignore
@@ -59,3 +59,5 @@ typings/
 
 # gulp build folder
 build/
+
+.DS_Store
