+++ package.json
@@ -4,7 +4,7 @@
   "description": "A Document management System",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon ./server.js --exec babel-node",
+    "start": "nodemon ./server/server.js --exec babel-node",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
