+++ server/controllers/users.js
@@ -1,7 +1,7 @@
 import bcrypt from 'bcrypt';
 import localStorage from 'local-storage';
 import jwt from 'jsonwebtoken';
-import jwtoken from '../../jwt';
+import jwtoken from '../helper/jwt';
 import Utils from '../helper/utils';
 
 require('dotenv').config();
