+++ server/route.js
@@ -8,7 +8,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
 // Require our routes into the application.
-require('./server/routes')(app);
+require('./routes')(app);
 
 app.get('*', (req, res) => res.status(200).send({
   message: 'INVALID ROUTE!!!.',
